Get it working, except FluxInspector
